ResponseHeaders class added. httpParseHeaders moved to Response Headers and splitted for testing.
