+++ Tests/Functional/Services/CurlTest.php
@@ -102,54 +102,4 @@ class CurlTest extends \PHPUnit_Framework_TestCase {
         $this->assertInstanceOf('Symfony\Component\HttpFoundation\Response', $response);
         $this->assertSame('application/json', $response->headers->get('Content-Type'));
     }
-    
-    /**
-     * @test
-     * @covers ::httpParseHeaders
-     * @covers ::<private>
-     */
-    public function testHttpParseHeaders() {
-        $headers = "HTTP/1.1 302 Found\r\n".
-            "Date: Tue, 21 Apr 2015 13:57:48 GMT\r\n".
-            "Server: Apache/2.2.22 (Debian)\r\n".
-            "Location: https://test.de\r\n".
-            "Vary: Accept-Encoding\r\n".
-            "Content-Length: 284\r\n".
-            "Content-Type: text/html; charset=iso-8859-1\r\n";
-        
-        $parsedHeader = Curl::httpParseHeaders($headers);
-        $this->assertArrayHasKey('Date', $parsedHeader);
-        $this->assertArrayHasKey('Server', $parsedHeader);
-        $this->assertArrayHasKey('Location', $parsedHeader);
-        $this->assertArrayHasKey('Vary', $parsedHeader);
-        $this->assertArrayHasKey('Content-Length', $parsedHeader);
-        $this->assertArrayHasKey('Content-Type', $parsedHeader);
-        $this->assertEquals("Tue, 21 Apr 2015 13:57:48 GMT", $parsedHeader["Date"]);
-        $this->assertEquals("Apache/2.2.22 (Debian)", $parsedHeader["Server"]);
-        $this->assertEquals("https://test.de", $parsedHeader["Location"]);
-        $this->assertEquals("Accept-Encoding", $parsedHeader["Vary"]);
-        $this->assertEquals("284", $parsedHeader["Content-Length"]);
-        $this->assertEquals("text/html; charset=iso-8859-1", $parsedHeader["Content-Type"]);
-        
-        $headers = "content-type: text/html; charset=UTF-8\r\n".
-            "Server: Funky/1.0\r\n".
-            "Set-Cookie: foo=bar\r\n".
-            "Set-Cookie: baz=quux\r\n".
-            "Folded: works\r\n\ttoo\r\n";
-        $returnValue = Curl::httpParseHeaders($headers);
-        $this->assertTrue(is_array($returnValue));
-        
-        $this->assertTrue(isset($returnValue['Content-Type']));
-        $this->assertEquals($returnValue['Content-Type'], "text/html; charset=UTF-8");
-        
-        $this->assertTrue(isset($returnValue['Server']));
-        $this->assertEquals($returnValue['Server'], "Funky/1.0");
-        
-        $this->assertTrue(isset($returnValue['Set-Cookie']));
-        $this->assertTrue(is_array($returnValue['Set-Cookie']));
-        $this->assertEquals(count($returnValue['Set-Cookie']), 2);
-        
-        $this->assertTrue(isset($returnValue['Folded']));
-        $this->assertEquals($returnValue['Folded'], "works too");
-    }
 }
